container-qvector: Speedup qv_sort using thread jobs.
This simple patch uses thread jobs in qv_sort in order to optimize it.

It is done only if the thr module is loaded, and if the vector is large
enough (I've chosen an arbitrary limit of 1024 elements).

On my simple benches (using the qvector.qv_sort test with big values, on
my 4 CPU machine),
I've observed a speedup from 2x to 3.5x, depending on the number of
elements.

Change-Id: Ib83d4cc84c5be5a5748d543380c6ac011549030d