core-mem-ring: use e_trace only if log module is loaded
As seen in #54184, if the log module is shutdown before the
ring-pool one, and there are leaks, a call to e_trace is done
, accessing freed memory in the log module.

A dependency cannot be explicited to the log module, as it
uses the mem pools itself. To avoid the issue, an explicit test
about the log module is thus done, and if not loaded, a simple
printf is done.

This requires making the log module public.

Closes: #54184
Change-Id: I0b11fb4dd57cbb49f30d84f8f825c6adc908aa89