iopc: Remove backward compatible versions and options.
Now that the iopc used to compile the repository is the one in the
repository, it makes no senseto maintain the backward compatibility
versions (v2 to v6). This commit properly deletes them.

Other options that were activated with v6 were also deleted.

WARNING: a consequence is that we should NEVER release in the tools an
iopc after this commit (because this would break the compilation on the
branches < 2017.2). Maintenance on these versions should be done on 2017.2
or 2017.3.

Change-Id: I015fa21055a9cd5fa2ae583956e533e80b2f7710