iopc: fix wrong split when a doxygen tag is found
When parsing the following doc, iopc put the part before \ref in brief and
the following part in details.

 /** Position of the field to use in the \ref Extract::fields.
  */
 uint pos;

With an explicit \brief keyword at the beginning, it would work, because
in that case the chunks are merged. I don't see why we require an explicit
keyword to merge, so I just removed that requirement, and it seems to work
properly.

Change-Id: Ifb176f9416ed5feccbf6e927b621d906b80a987f
Closes: #54939