iop[backward-compat]: Be stricter with non-strict enums.
Before this patch, changing the numeric value of an enumeration member was
not considered backward-incompatible if the enum was not strict (because
this does not provoke unpacking errors).

But this was really dangerous because it allowed to break an enumeration
of event.iop (which is the main use-case of this lib) without any error,
because enums of event.iop are not strict.

To fix that and make it simple, just forbid it even for non-strict enums.
Also detect when two values are inverted.

Change-Id: I55867c4ca2aece8ac8aeb9fbb5536bf9ca4e3eea
Closes: #55117