iop/env: implement unified getter for struct/union/enum/class
The getter for union/struct was missing. Add an unified getter
"iop_get_obj()" and get rid of the hash tables "classes_by_name" and
"enums_by_fullname".

Use an unified hash table for a new unified IOP object 'iop_obj_t' that
can be a package, an enum, a class, a struct or an union, since we don't
expect to have name collisions (excepted the ones created by DSO loading).

Change-Id: Ia4dc1aa59165040a310e18467f989f02887eb8a0