iop: fix return type for iop_bpack_size (the function cannot fail)
The function iop_bpack_size had the same return type as
iop_bpack_size_flags but this last one could return an error only when
using the flag IOP_BPACK_STRICT, which was not used in iop_bpack_size
implementation.

This patch switches the return type to an unsigned one so the caller can
know that he does not have to handle the error.

Change-Id: I788285284460abf16f4764678f79e554d9aa3e20