thr: Change default running_on value for queues.
Problem:
~~~~~~~~
thr_is_on_queue(thr_queue_main_g) could return true when called from
threads that are not thread jobs (like the zookeeper thread), which is of
course wrong and could lead to unexpected behaviors.

This is because this condition could be true when the queue was not
currently running (running_on == -1):

  atomic_load(&q->running_on) == (ssize_t)thr_id()

Because thr_id() is -1 on non thread-jobs threads.

Solution:
~~~~~~~~~
Change the default running_on value to -2!

Change-Id: I11922aee7eb34007b1e5c004a0b7f5c313420db5
Closes: #56405