el: fix buggy heap reordering after rearming repeated timers.
In case we have a repeated timer, we rearm the timer after its execution,
then reorder the heap in order to move the timer to its right position.

However, when reordering the heap, we supposed that the current timer was
at the top of the heap, which may be untrue in case a new timer has been
registered during the execution of the callback of the timer. As a
consequence, we may have corrupted the heap in that case.

This may be responsible for the bug observed in #47754 were some
fast-paced timers are not at all fired during a whole event loop (but it's
hard to say if this is actually the fix for it).

Change-Id: Ibb7a6e008bd1452028b0d9b94c6ae8b4c9c8236e
Refs: #47754
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>