mem-stack: Change reset policy.
Problem:
~~~~~~~~
Before this patch, mem-stack pools were reset every UINT16_MAX (65535)
calls to pop() (once it is empty).

The problem is that in practice, this was really rare, especially in
stacks that were mostly idle (which happens the t_pool of the inactive
thread jobs for example, or even in the main thread t_pool in a daemon
that does quite nothing).

This could lead to memory never released, which becomes a real problem in
case there are a lot of threads.

Solution:
~~~~~~~~~
Change the policy to reset the stack pools when their size exceeds a
certain threshold (10 MiB for the main thread, 1 MiB for the others), and
at most every minute.

This is still done only when calling pop(). The following patch will
ensure this is checked on a regular basis for all the t_pools of each
process.

Change-Id: Ic3acf9af6cdf83e4fefd2907bc6ac96e9de40c95
Refs: #56342