Fix core mem modules registration.
Problem:
~~~~~~~~
The core_mem_* modules were declared as NEEDED_BY el. The problem is that
the el module is required in a constructor, so depending on the
constructors order, the el module could be running when registering a
core_mem one, leading to an assert.

Solution:
~~~~~~~~~
Don't automatically require core_mem modules, and consider this will be up
to the programs to require them (a core_mem module was added for that).

I could have declared functions to ensure core_mem modules are registered
in the constructor of the el module, but this was too complicated.

Change-Id: I03e8e8b39c647d32fde57b1187f6fd6de6cc3fb9
Refs: #56342