core-mem-libc: Call malloc_trim in a cron.
Our tests show that some daemons of our products seem to consume a lot of
memory when doing nothing because malloc does not release memory to the
kernel (observed at least with QTS).

Calling malloc_trim solves this issue. Since there is not a real good
moment to call it, do it in a cron.

Change-Id: Iaa3915eafc0ee7e690f1ec825f40de7d6c8914cf
Refs: #56342