qps: consume snapshot process events.
While qps waits for the child snapshot process to terminate, the child may
send logs, using ic_flush(). If the ichannel is not read by the parent,
the child will block when a limit is reached, creating a deadlock.

To fix this, we call a module method while we wait, which will take care
of reading the child events.

Refs: #56311
Change-Id: I9b76b6ef1d4818e5cbc005779813d20a317dc0cc