core-mem: add functions to set the malloc trim period and pad
The change Iaa3915eafc0ee7e690f1ec825f40de7d6c8914cf added a cron that
calls malloc_trim in order to force malloc to give back memory to the
system once in a while, but for some use cases the default period of 10
minutes is not appropriate, for example if we want to avoid some memory
use peak in big calculations.

This patch adds two functions to change the period and the pad parameter
to use when calling "malloc_trim()".

Refs: #56342
Change-Id: I34eb6c3c9ba22ff0d23a86d0ffe43bf9d534de8d