core-mem: Unref malloc_trim_cron after its registration.
I initially choose to unref it in the on_term only so that we can see it
in the el blocking summary (printed when receiving a KILL -PWR) for debug
purposes.

This this is an issue for some pieces of code that do not call
module_on_term (like qmr-sc, which makes the tests fail).

Change-Id: I259cbc276f8dbcb31bfcc7e6ae632969a05ac2b3