iop-rpc-channel: shorten bytes fields for ic tracing feature
Before this patch, when tracing (with the ic tracing feature) an RPC that
contains a byte field (as input or output), there was a risk to be
completely flooded if the amount of data was important. This patch solves
this problem by simply keeping a few characters for byte fields, while
other fields are left untouched.

Change-Id: I27e3ad78ceef2d95792c714be2e0d65b3c7fcd47