iop-hash: take class_id into account when hashing a class
The class_id was not taken into account, leading to

 {
     "_class": "pkg.Foo",
     "int1": 1
 }

and

 {
     "_class": "pkg.Bar",
     "int1": 1
 }

to have the same hash. This is fixed by adding the class_id of the
instance in the hash.

/!\ This breaks licence signatures, and will require re-signing every
licence, as it adds the class_id of every licence module in the hash.
It is pushed in master to avoid breaking existing installed products.

Change-Id: I903a33f886211cbd2da6fa135e6dfc16fd0873c2