log: Make log buffers thread safe!
It seems crazy, but there is a huge bug in the log buffers from the
beginning: they are not thread safe! They are indeed using global
variables to handle their states (mem stack pool and stack of buffer
instances).

To fix that, simply move the log buffer related variables in log_thr_g,
which is thread local.

Change-Id: If72ba03a0fddb028a0a7b4b4ede8203a228c4c6e
Closes: #57316