iop-filter: Add bitmap variants.
This introduces the iop_filter_bitmap and iop_filter_opt_bitmap that do
not modify the input vector but fills a presence bitmap according to the
filters.
iop_filter_bitmap_apply can then be used to apply the filters on the
vector.

This will be used in platform to apply more complex conditions, like OR on
different filters.

Change-Id: I02b465fd28cc459ebe483afb67f7a7d740e2903a
Refs: #57303