ichannel: try to flush answers immediately even if POLLOUT wasn't requested
6b1d8f6ff1ad82aaa57af7b6daaf7ceffdd11812 tried to write queries answers
as soon as possible to improve the reactivity of ichannels. However, we
tried to flush answers only if the POLLOUT flag was already set. Which
meant we only improved the reactivity when some queries were already
pending to be written.

To fix this, we don't even consider POLLOUT in ic_read(), we just try to
write and if it fails we won't try again, that's it.

Refs: #57072
Change-Id: I88d2c51fd277806f53399f855f885a5bc7130b1f