iop-hash: parametrize taking class_id into account or not
In ecc85383eef32bbd the iop hashing algo was updated to take the class_id
into account for classes. This is needed and known to break the backward
compatibility but what was forgotten is that we also need to keep the old
behavior in order to be able to sign licences for old and already released
product versions.
The parameter allows skipping taking class_id into account so that only a
few code has to use it, since most of the time the behavior before this
patch is the one we want.

Change-Id: I29f2880011203bfc2f34a129fbb47f96904be6a1