web check: Don't skip web init when Z_TAG_OR is set
When we use Z_TAG_OR to launch a bunch of tests, the selenium tests could
not be run. This is because the init (setting the browser, the server
address etc) was not correctly done because the "web" was in the
Z_TAG_SKIP env var. As Z_TAG_OR override Z_TAG_SKIP we now check
that the Z_TAG_OR is also empty to safely skip the web init, if Z_TAG_OR
is set we better init the web (it's actually quite harmless to do so)

Change-Id: I7b2a7c5bc9067dcbcae0ce3700a63f828e9b8716
Closes: #48004