el: remove constraint on first argv in el_spawn_child
As stated in execvp man, by convention the first argument should be the
same as the file being executed.
Before this patch, this constraint wasn't checked nor handled for the user
by el_spawn_child. Now the user only has to provide the additional
arguments and el_spawn_child concatenates those arguments with the name of
the file being executed.

Change-Id: I4a41e44d718aa55cdda8cf0f15dd6d550d146b35