iopc: fix struct packing (C fields ordering) with mandatory class fields
When packing C structures (reordering fields to avoid holes), IOPC didn't
make a difference between IOP structs and IOP classes, but the second are
pointers when the first are not.

This issue led to wrong size and alignment values for mandatory class
fields, though the impact was not important because it was not present in
generated IOP structures (the size and alignment is generated with
'sizeof' and 'alignof').

Change-Id: I1febbf0a61850278dfef31ea4b183d9f544a95a4