run_checks: remove unset_www options from run_checks
This commit partially reverts
292a8d4717908c9da42defdd3fa0886210a191db
that introduced a buggy remove_www option that is removing user-specific
options when running tests using make.

+ Bug +

In this loop, we are modifying/exporting variables that are evaluated at
each run. The issue here is that if the loop runs multiple times, we
evaluate the first time against user-defined options, while the next time
on automatic set variables. The bugged approach used before is forget
everything each loop run, which is wrong because we also forget user's
defined options.

+ Fix +

We now save user defined variables, and we evaluate them on each run. If
not set, we export some automatic defined variables and we export them.

Change-Id: Iebdf86a447921e29326b0fd23c134172c2487097