arith-float: add double_round_significant
This new method is similar to double_round but takes into account the
number of significant digits of the integer part of the number.

The goal is to round the number to a number of significant digits.
However, if the integer part has more significant digits than is
requested, the number is kept as is, to avoid issues with non normal
values, or too big values, etc.

This will be used in QKV to convert stored float values to double values
with a certain precision.

Change-Id: I5d78e13ae6dca8a2732c350a82fa19453dbe97f1
Refs: #56617