iop: adapt "iop_dup" code to refactor "mp_iop_dup_desc_multi_alloc"
The function "mp_iop_dup_desc_multi_alloc" allows to duplicate an IOP
structure with several allocations instead of a big single one, provided
that the memory pool is frame-based and that the user doesn't expect to
get the size of the allocated block.

The function was originally implemented in 2017.2 from a copy of
"mp_iop_dup_desc_sz" functions taken from master. The code was duplicated
in order to avoid changing anything in the original function in stable
branches.

The function "mp_iop_dup_desc_sz" allocated the content of duplicated
structures within a single allocated block, to do so, it has to scan the
struct twice: one to calculate the allocation size (iop_dup_size), another
to copy the content (__mp_iop_dup_desc_sz).

The main interest of doing so is to give the capability to the user to
deallocate it with a single "mp_delete" call.

The interest is more limited when it comes to "frame-based" memory pools
(stack pool or ring pool), that do not have an implementation for "delete"
because they don't work this way, so we still get the double scan overhead
without using the advantage is supposed to bring. We also have a big risk
of reaching MEM_ALLOC_MAX with a big single memory allocation.

The new implementation could become the new default for frame-based pools
but benches showed that the CPU time gained with the suppression of the
call to 'iop_dup_size' didn't always compensate the allocation overhead
(in some pathological cases) so we need some deeper benches first.

Refs: #57751
Change-Id: I809dabaa9c404b37b8543e9682664740e0a0edf6