iop²: implement a test based on "tstiop.FullStruct"
The struct "tstiop.FullStruct" is supposed to contain any existing field
type in IOPs. Contrarily to other test structures, this one is supposed to
be shared between the tests that could need such a complete structure.

Use it to check IOP² structs generation.

Note: the development of this struct allowed us to see some still
undetected bugs.

Change-Id: I855fee44e94abcd721d446f150dbad872473dfb5
Refs: #57751