iopc/lang-c: constness + style fixes
The "const" qualifier was used in a random way in the file. It is still
missing in some cases in which the functions "pretty_path" and
"pretty_path_dots" are used on a struct attribute because those functions
modify the "iopc_path_t" we give to them.

Also remove unexpected uses of "uint" type (use "unsigned" instead) and
some coding rules violation (not all of them).

Change-Id: I04103737c2a990ca5295977613cc360792cb6845