iop²: detect tag conflict in IOP structures
Make the field builder check that a tag is not used twice in a given
structure. Also add a Z-TEST about it.

Note: perform a linear search to search for an already used tag just like
we do in iopc parser. It's not so expensive with a reasonable amount of
fields and it also keeps the code simple, but maybe we should considerate
switch to a more efficient check.

Change-Id: I6877d7d92ce59f1fdbcc0ccf1c4cc724cd885b94
Refs: #55751