iop²: fix and test IOP array field case
Fix1: the "repeat" attribute was always overwritten by
      "iopc_field_set_opt_info", that won't happen anymore.

Fix2: the field "size" in "iop_field_t" was wrong for arrays: the right
      size to use was the size of the "required" version of the field.

In zchk: rework the checks that compare the generated "iop_struct_t" to a
reference struct generated with the IOPC compiler.

Note: more complete tests should be added.

Refs: #57747
Change-Id: Idb927c652d2be6161529b72b6d7ff1933f8ab9e4