iop²: introduce 'iopsq_type_table_t'
When using IOP² to generate struct descriptions, we may have to create
structs with field that also have generated types.

The only thing to do is to register the generated types we need in an IOP²
type table so it can be referenced with an identifier (the new field type
"typeId"), then pass the type table to the API when generating the IOP
description.

Change-Id: Iebd862187fcb7f47f579cead76c325f44de9374f