build: fix race when generating core-version.c
The core-version.c used to be updated by the `update-if-changed` macro.
However this macro is racy when two parallel targets depended on
core-version.c and tried to generate it at the same time. During the
removal of the temporary generated file of the first run, the second run
could consider core-version.c as different and then force a rebuild.

So we now use a dedicated function, `gen-and-update-if-changed`, which
uses a unique temporary filename, suppressing any race during parallel
executions.

Change-Id: I091ecfba115b95f720a1046f437173a1e9f18d8a
(cherry picked from commit ba5602fde816b907488857e8993b3fca2e6a535a)