iop.swift: IOPClass make method without env parameter.
/!\: Since env is optional in the methods, this new method *should not* be
needed at all. However, it seems that for optional classes, swift fails to
resolve those to the method with an additional defaulted parameter.
We therefore add a stupid method that performs the defaulting "manually".

This seems to be a strait up swift compiler bug. Could not find a report
about it (and since swift4 is out, it is probably already fixed).

Change-Id: I31467f6e137fb30095bf5a6ecdd5cdd8bc808e97
Closes: #58489