xmlr: fix missing call to xmlr_fail.
When we exit any function of the XML reader library with an error (i.e a
negative value), the error description in xmlr_err_g should always be
set to indicate to the caller where the parsing failed.

This is done by:
- xmlr_err_handler for the errors that occur in the function of libxml2
- a call to xmlr_fail when the errors occur in our code.

This patch fixes several cases where we triggered an error without
calling xmlr_fail. This produced a later crash since we were not
expecting the pointer to the description to be NULL.

Change-Id: I4b9323841cfa735dfe9c9ce69114a58f26cec864
Refs: #58907