Revert "iopc: Drop swift support."
This reverts commit f65ce42347561b6c7c4f9eb34ab3f2bf7a655843.

The problem is that 2017.3 is the branch in which iopc is released (in the
tools), for the branches that do not use the committed iopc in their
buildchain.

So dropping the swift support in the released iopc would break the
compilation on 2017.2, which uses (well, used) swift and the iopc from the
tools.

swift support from iopc will be really dropped in 2017.4.

Change-Id: I7719dd10cac2834c91a05e4ea9c2fa521bb713d6