build[farch]: Generate .h files, not .c.
Before this patch, the build system was generating .c files when
processing a .fc file using farch. This .c file was then included in
commited C sources.

This makes the integration of other build systems more complicated because
they tend to compile the generated c file, while it should not be
compiled itself.
Generating a header fixes this issue.

Change-Id: Ia509aee613937f45e4d2c397859341d1ef2020f6