core-mem: Explictly depend on el in core_mem_libc.
Both core_mem_libc and el are required in a constructor and released in a
destructor.

Before this patch, there was no relation between both, so core_mem_libc
could be shutdown after el, which is invalid because
core_mem_libc_shutdown unregisters an event.

This caused asserts in the el module at shutdown:

 el_shutdown: {el} 1 events are leaked...
 el.blk:329: el_destroy: Assertion `module_is_loaded(el_module)
                                 || module_is_shutting_down(el_module)'
 failed.

Change-Id: Ic1d97ebc5a68ce8971fe9c1b55006f2478d991ef