farch: obfuscate file names and contents.
This patch does the following:
  - cut file contents in randomly small symbols,
  - obfuscate file names and contents.

It introduces the following changes:
  - the name of an entry is now a lstr (it may contains '\0'),
  - entries and their contents are separated in two arrays:
    * the first array (called "<archive_name>_data") stores the contents
      of all files of the archive,
    * the second array (called "<archive_name>" as before) stores the
      archive entries,
  - file names and contents are obfuscated by xoring them with already
    existing symbols copied by hand in an array.

It also embed a hack preventing that two simultaneous execution of farchc on
the same file is possible without races. The bug, which belongs to our build
system, was present before this patch. However, it was not a problem because
the content of the file was deterministic.

Refs: #59195
Change-Id: Ic7acd2fedf27d6a30be0ddaab33679f0dbaf5897