First version of waf build system.
Here it is, the lib-common (almost) builds entirely with waf! Thanks to
Pierre P for its initial work.

To compile:
* Install waf (https://waf.io/)
* waf configure build

As you will see, this is reaaaaaaally fast: ~10.5s against 22s for make
P=debug NOCHECK=1!

There are still some hacks and a lot of stuff to fix (TODOs and FIXMEs
left in the code).

Change-Id: I31b5c82ea4a40b8b21d736cb469d5ae6cc3e7b0e