iop: implement iop_xsort/iop_xpsort/iop_cmp
These new functions allow to sort an array of IOPs following an arbitrary
comparison operator.

Purpose:
~~~~~~~~
- Simplify IOP sorting when:
    - We want the sort to be deterministic (all the fields are
      considered).
    - The IOP are generated internally, so we can choose the order of the
      fields in function of the sorting order we want to apply.
- Have a sorting function for testing the output of a function that
  generates unordered IOPs output.

The code is a rework of "iop_equals()" code, and the zchk C is the same.

Note: some TODOs have been left for a flag that would unlock quicker
comparisons instead of natural ones, implementing such a flag would
mean:
    - Compare the union tags number instead of field names
    - Compare class IDs instead of class names
    - Compare arrays content only if the length matches
    - Use "lstr_cmp" instead of "lstr_utf8_icmp"
    - (optional) Compare full arrays of integers with memcmp instead of
      comparing element by element.

The flag mentioned by the TODOs may be implemented for cases in which we
don't need a "natural" sort (xsort tries to be as natural as possible).

Change-Id: I62c0a4c568634912d5e60c5d32d7f11cdc700abb