waf: set `iopc` dependency to build IOP files
Building an IOP from a subdirectory calls `process_iop` but it does not make
sure `iopc` has been built before. The dependency was not specified and we
were relying on the target build order.

When the task generator of `iopc` is initialized we post the task so that we
can get the `link_task` which builds the executable. It is then used by
`set_run_after` which set the dependency.

Change-Id: Ib75dd82c323469added2ed097bd97aeb3323d259