libcommon: fix getppid for Docker.
In the from_editor function, that seems to check if the compilation is
launched from emacs or vim, a loop was looking recursively for the ppid of
the current executable, and was supposed to end when reaching "1". But in
Docker, we quickly arrive to the "0" ppid, and the loop was then
continuing with random values found from the command "cat /proc//stat",
since the ppid of 0 is null, and it was failing continuously. So, we now exit
the loop when the ppid is smaller than 1.

Change-Id: Ic6ab07a1a75f90261f8dadd234a3844819f0f44a