waf: Get rid of warnings in 'waf -v'.
I'm not 100% sure what the exact reason was, but the code_generation group
was responsible for tasks posted in double (once in the core_generation
group, and once in the code_compiling one).

After some tests, it finally seem that this group was useless and that the
two other ones are enough.

Change-Id: Iab6d80fb0ddd4e3f3f4e1eb6c974acd0d713aa53