iop: add NO_REALLOC flags for iop_copy.
Do not perform reallocation of the output value on copy.

For class instances, the output object must be initialized and the
original __vptr of the output object is not modified.

Like IOP_COPY_MULTIPLE_ALLOC, the memory pool must be a by-frame
memory pool if IOP_COPY_SHALLOW is not used.

This flag is not available for iop_dup functions.

Change-Id: I27774093fd891d3f5a73c96456f9bc80a8a85915