farch: simplify obfuscation and data structures.
In farch.h, we used a random xor_key for data chunks and filenames.
Tough, a chunk size is randomized: it can be used as chunk's xor_key.
Similarly, the number of chunks has no reason to be the same between
two farch archives and can be used as xor_key.

Without its xor_key, a farch_data_t is just a lstr_t. We loose the
"const" modifier, but the code is much simpler.

I also add a few lines (in farchc) to format long strings as we would
do in real code.

Refs: #59589
Change-Id: Ic5461cbcddca144a689a3e1f76d9c27e25877fd8