iop-TS: fix optional repeated fields.
When instantiating IOP models, defining repeated fields should not be
mandatory.

Example: with IOP

    struct Labels {
      string? en;
      string? fr;
      AdditionalLabel[] additionalLabels;
    }

We should be able to write:

    new Labels({ en: 'foo' });

It was not possible because repeated fields were considered as mandatory
in TS IOP definition.

Now, when generating the TS definition from IOP, repeated fields are set
as optional for ModelParam. So it lets us declaring model without setting
array.

Repeated fields in Model are still considered as mandatory, so we can use
these fields without checking they exist.

Change-Id: I4f27c4c2eea1e72b5aa2bf31c15f32f96e1f134e