build: add ale target to build clang flags for ALE vim plugin
(cherry-pick from e0bbe71d69e94261604b6381e73309c58f77c44a)

For the ALE plugin, in order to have asynchronous linting for C files,
the compilation flags must be provided.
The preferred way is through a local vimrc file at the root of the
project.

This commit adds a target to build this file.

Change-Id: I9ac1e5bb7d3f2044c20406615d7cb919b340d4c6