waf: Change the way iopc options are handled.
This commit introduces the IopcOptions class, that allows to register the
options given to iopc for the IOP files of a given path (which allows
mimicking the Make behavior).

The main point of doing that is to recursively evaluate the includes,
which is necessary for building the IOPs in mmsx.

Change-Id: I31e2f0380a1a57dc2ee8900f37c53a0fe570baa7