waf: Ensure farchc/iopc build order.
iopc uses a farch file in its sources, and before this patch there was no
guarantee on the build order between iopc and farchc.

It seems it was working thanks to luck, but small changes in the tasks
declarations could break it.

To solve that, use a dedicated group for farchc.

Change-Id: I25b7b2f8a3bfda18467f9a6744d14ee544d7c7e5