waf: Generate source files in the source directories.
waf, by default, generates all the source files (fc.h, iop.h, ...) in the
build directory, so that they are not mixed with the committed source
files.

This looks like a good idea, but has some big drawbacks:
* it requires doing a lot of hacks in inclusion paths so that it actually
  works, especially in mmsx.
* it will probably be really hard to make linters work in these
  conditions.

Make was generating source files in the source directory, so let's do the
same.

Change-Id: I1210bcad5d5aacf9a2b131ee1fac55fb7274f4df