waf: Fix 'use' includes in clang rewriter.
Before this patch, the include paths coming from 'use' attribute of task
generators were totally ignored when calling clang, making it impossible
to compile some targets in mmsx.

The solution for now is to manually extract the '-I' flags from the
original task environment to set them in the generated task environment.

As explained in the code, the solution is not perfect for now, but at
least it works fine.

Change-Id: I1c1be77b2ff88cef5e4fdf5924e800b82a487731