cflags.sh: fix from_editor, do not write empty "1" file
">" is redirection, not a comparison operator for the "test" command.

By a comical string of events, this caused selenium tests to fail in
One when trying to write "1" in inputs: the send_keys("1") command in
selenium actually tries to detect a file by this name, and then replace
the value by a upload path of the file. Top quality code by selenium.
Additionally, this only happened in remote mode. Why? I haven't
investigated so probably some other black magic from selenium.

Change-Id: I92b5b832fbfb0c8b3189b51be50a13d69d095ba5