iop: add helpers to wrap an IOP class in a core object
In many cases, we want to associate core objects classes (or C classes) to
our IOP classes in order to associate C code to the IOP classes.

It usually implies some administrivia: association table between C classes
and IOP classes and registration API.

This patches adds macros and functions to ease this process by providing
the macros to create the functions for registration and construction of
such objects.

Change-Id: I4e58d7b96aef41d03a5d50f4e9b3c15a8617837b