qhash/qhugehash: protect *for_each* helpers against pos corruption
When using "tab_for_each_pos", it can be considered as obvious that the
"pos" should not be changed during the loop because the scan is trivial
and the programmer guesses how it works.

It may be not as obvious to the programmer that it is the same for
"{qh,qm,qhh,qhm}_for_each_pos": the "pos" may be understood as a
mutable by-product of a complex iterating process, but it is actually
necessary to keep it untouched if we want to get to the next element.

This patch protects the qhash and qhugehash scanning helpers against
such corruptions by keeping a private copy of the position so the
position variable can be altered without consequence. It fixes issues
like the one fixed in I3b180be8a4cfe40e3265b5f06738f004fd6c30ca.

Change-Id: Idb5e69dce58f9b0724bd924a5214598f475252b4