iop: avoid useless iop_init_fields when calling iop_init_desc on unions
I found a case where iop_init was called on a union type
(platform.DataType) which is a bit stupid, but what iop_init_desc did was
also stupid since it was initializing all fields of the union, one by one.

Note that the recursion in iop_init_fields is properly done : only for
mandatory structs, not for unions.

Change-Id: I91a5989ea7e2c98cd2d514f0be2654959449e56e