ssl: encrypt and decrypt arbitrarily long strings with RSA keys.
RSA keys can only encrypt data as long as the size of the key. A 2048
bit key can only encrypt 256 bytes. Plus, the encryption should (to
increase security) add a random padding. While the most common padding
weight 11 bytes, the new recommanded one is about 41 bytes, which
reduces again the available space for encryption.

Plus, it is NOT RECOMMANDED to split a string and encode each part
with the RSA key (especially for performance purposes). Instead, we
generate a symmetric AES key and a random initialisation vector, and
we put them together with all data that can fit in the RSA header. The
rest of the data are encrypted with the AES key.

  [ -------- RSA encrypted data --------- ][ AES encrypted data ]
  [ AES key | AES IV | Data | RSA padding ||    rest of data    ]

However, for small strings (ligher than rsa_size - rsa_padding bytes),
all data can fit in one RSA header. Thus we continue to use only this
header.

  [ -------- RSA encrypted data --------- ]
  [          Data           | RSA padding ]

Refs: #60793
Change-Id: Id2f20026396c3bc9b2756f1ccd132e97192aacbd