farch: Do not support wildcards anymore.
Problem:
~~~~~~~~
waf does not seem to support wildcards as output of the scan method. As a
result, when adding a new file that matches the pattern, the .fc.h is not
rebuilt as it was with Make.

Solution:
~~~~~~~~~
Just don't support wildcards anymore. This simplifies the code of farchc,
and will make the farch files more explicit, which is not a bad thing...

Change-Id: Ic3462cbfd6e5867ad69f741a0942f3b31564b9de