iopc: Dump "deprecated" attribute in iop.json files.
Before this patch, the information that a field is marked as @deprecated
was not present in the .iop.json files.

Add this information, because it will be very welcome for handling
iop-form translations (to skip the deprecated fields). This could also be
used by iopdocgen when generating the .adoc files.

Change-Id: I12b7bb1a8ed70c72ccb904dc9e39d9e0a0a39cc1
Refs: #61459