build/www: have bundles in "src" directory, not "node_modules"
node_modules is really for installed node packages, not for files to
compile. This is an issue for webpack which refuses to compile
typescript files in a node_modules directory.

Rename node_modules to src. To keep the compilation of one modules
working, the declarations files still need to be copied in
.build-XXX/node_modules.

Change-Id: Iba1b69a701beb027bb58449d452d37855a21e9b6