iop-jpack: allow dumping same content in the same subfile twice.
In some cases, for configurations, we want to include the same subfile
twice in the same structure.

Before, on packing, we simply dumped the same subfile twice and we
didn't actually checked that the content was the same.
It returned an error if the subfile is being created with a read only
mode.

Now, we check if the subfile has been already written and the content is
strictly the same.

iop_type_is_string() and iop_value_equals() needed to be moved from
iop.blk to iop-helpers.in.c because we are now using them in
iop-json.blk.

Change-Id: Id59392da0a79faa1f5ee8583307d830b513b1ba9
Closes: #61865