iop-jpack: fix writing string subfile flags and mode.
Before, when writing a string subfile on jsong packing, we used the
flags and mode by default of xwrite_file instead of the one in the
context.

Now, we correctly use the ones specified.

This required modifying and exporting file_flags_to_open_flags() and
xwrite_file_ex() to specify the flags and mode used when writing.

Change-Id: Iba8b8d6979551135e938a406ad09b6415b43df83
Refs: #61865