iop-rpc-channel: fix SSL rbio swap for TLS handshake for OpenSSL 1.0.1.
Old version of OpenSSL does not support swapping the read BIO of the SSL
object in middle of the TLS handshake like we did before.

However we still need to it since we need to read the remaining data of
the buffer before actually using the socket.

The solution is to use a null filter BIO as the read BIO of the SSL
object and swap the next BIO when necessary.

Also use BIO_set_mem_eof_return for the buffer BIO so we it will return
a SSL_ERROR_WANT_READ when all the buffer has been read same as the
socket.

Defining SSL_set0_rbio ourselves is no longer necessary since we only
set the BIO of the SSL object once.

Change-Id: Idb976f5981f11bb46f560d3660fd45959b4f6b87
Refs: #60793