build[www]: Produce (and handle) read-only css.map files.
Problem:
~~~~~~~~
waf produces read-only files (.css and .css.map) when processing a less
file, but the .css.map file produced by Make was not read-only (only the
.css was).

The problem is that Make failed when a read-only .css.map file was present
on disk (so running waf and then make failed), because it was not designed
to handle them (cf .89cd50d846).

Solution:
~~~~~~~~~
Produce (and thus handle) read-only .css.map files.

Change-Id: I26527edb96b116d391f526aac96904c547b912bc