Build: allow declaring webpack bundles used for www-check
Add _WEBPACKTESTS used to declare a webpack bundle used for jasmine
tests.
However, as nothing is simple with this superb build system, we cannot
just add it as a dependency of "www-check" (that would of course start
the bundling in parallel of the tests that need it).

To be able to have this dependency and *wait* for it to be completed
before executing the tests, a phony "www-check-deps" is added.

Change-Id: I07039684aa049b979b41872104ef555aee0f15ea