waf/common: fix uid of CustomInstall tasks.
All waf tasks must have a unique id that is by default computed from the
class name, inputs and outputs.
Since CustomInstall don't have any input and output, the uid was always
the same which is bad.
Now, we compute an unique id from the class name and the commands.

Change-Id: I8a102f561004c032d4c27ded85388c23ea843547